[pkg/logs] Move decoder, metrics, pb, processor and tag to `internal`
