+++ pkg/logs/internal/decoder/auto_multiline_handler.go
