+++ pkg/logs/internal/decoder/breaker/line_breaker.go
