+++ pkg/logs/internal/decoder/breaker/line_breaker_test.go
