+++ pkg/logs/internal/decoder/line_handler.go
