+++ pkg/logs/internal/decoder/line_handler_benchmark_test.go
