+++ pkg/logs/internal/decoder/line_parser_test.go
