+++ pkg/logs/internal/decoder/single_line_handler.go
