+++ pkg/logs/internal/metrics/metrics.go
