+++ pkg/logs/internal/metrics/metrics_test.go
