+++ pkg/logs/internal/processor/encoder.go
