+++ pkg/logs/internal/processor/processor_test.go
