+++ pkg/logs/internal/processor/raw.go
