+++ pkg/logs/status/test_utils.go
@@ -7,7 +7,7 @@ package status
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
-	"github.com/DataDog/datadog-agent/pkg/logs/metrics"
+	"github.com/DataDog/datadog-agent/pkg/logs/internal/metrics"
 )
 
 // InitStatus initialize a status builder
